Prefactor: Move code into services layer and utils
