+++ hq_superset/utils.py
@@ -1,4 +1,4 @@
-import os
+import ast
 from contextlib import contextmanager
 from datetime import date, datetime
 from zipfile import ZipFile
@@ -6,14 +6,11 @@ from zipfile import ZipFile
 import pandas
 import sqlalchemy
 from flask_login import current_user
-from superset.extensions import cache_manager
 
 DOMAIN_PREFIX = "hqdomain_"
 SESSION_USER_DOMAINS_KEY = "user_hq_domains"
 SESSION_OAUTH_RESPONSE_KEY = "oauth_response"
 HQ_DB_CONNECTION_NAME = "HQ Data"
-
-ASYNC_DATASOURCE_IMPORT_LIMIT_IN_BYTES = 5_000_000  # ~5MB
 
 
 class CCHQApiException(Exception):
@@ -108,33 +105,6 @@ def parse_date(date_str):
         return date_str
 
 
-class AsyncImportHelper:
-    def __init__(self, domain, datasource_id):
-        self.domain = domain
-        self.datasource_id = datasource_id
-
-    @property
-    def progress_key(self):
-        return f"{self.domain}_{self.datasource_id}_import_task_id"
-
-    @property
-    def task_id(self):
-        return cache_manager.cache.get(self.progress_key)
-
-    def is_import_in_progress(self):
-        if not self.task_id:
-            return False
-        from celery.result import AsyncResult
-        res = AsyncResult(self.task_id)
-        return not res.ready()
-
-    def mark_as_in_progress(self, task_id):
-        cache_manager.cache.set(self.progress_key, task_id)
-
-    def mark_as_complete(self):
-        cache_manager.cache.delete(self.progress_key)
-
-
 class DomainSyncUtil:
 
     def __init__(self, security_manager):
@@ -191,24 +161,35 @@ def get_datasource_file(path):
         yield zipfile.open(filename)
 
 
-def download_datasource(provider, oauth_token, domain, datasource_id):
-    import superset
-    datasource_url = get_datasource_export_url(domain, datasource_id)
-    response = provider.get(datasource_url, token=oauth_token)
-    if response.status_code != 200:
-        raise CCHQApiException("Error downloading the UCR export from HQ")
+def convert_to_array(string_array):
+    """
+    Converts the string representation of a list to a list.
+    >>> convert_to_array("['hello', 'world']")
+    ['hello', 'world']
 
-    filename = f"{datasource_id}_{datetime.now()}.zip"
-    path = os.path.join(superset.config.SHARED_DIR, filename)
-    with open(path, "wb") as f:
-        f.write(response.content)
+    >>> convert_to_array("'hello', 'world'")
+    ['hello', 'world']
 
-    return path, len(response.content)
+    >>> convert_to_array("[None]")
+    []
 
+    >>> convert_to_array("hello, world")
+    []
+    """
 
-def get_datasource_defn(provider, oauth_token, domain, datasource_id):
-    url = get_datasource_details_url(domain, datasource_id)
-    response = provider.get(url, token=oauth_token)
-    if response.status_code != 200:
-        raise CCHQApiException("Error downloading the UCR definition from HQ")
-    return response.json()
+    def array_is_falsy(array_values):
+        return not array_values or array_values == [None]
+
+    try:
+        array_values = ast.literal_eval(string_array)
+    except ValueError:
+        return []
+
+    if isinstance(array_values, tuple):
+        array_values = list(array_values)
+
+    # Test for corner cases
+    if array_is_falsy(array_values):
+        return []
+
+    return array_values
