+++ hq_superset/views.py
@@ -1,14 +1,10 @@
-import ast
 import logging
 import os
 
-import pandas as pd
-import sqlalchemy
 import superset
 from flask import Response, abort, flash, g, redirect, request, url_for
 from flask_appbuilder import expose
 from flask_appbuilder.security.decorators import has_access, permission_name
-from sqlalchemy.dialects import postgresql
 from superset import db
 from superset.commands.dataset.delete import DeleteDatasetCommand
 from superset.commands.dataset.exceptions import (
@@ -17,26 +13,25 @@ from superset.commands.dataset.exceptions import (
     DatasetNotFoundError,
 )
 from superset.connectors.sqla.models import SqlaTable
-from superset.models.core import Database
-from superset.sql_parse import Table
 from superset.views.base import BaseSupersetView
 
 from .hq_domain import user_domains
 from .oauth import get_valid_cchq_oauth_token
+from .services import (
+    AsyncImportHelper,
+    download_datasource,
+    get_datasource_defn,
+    refresh_hq_datasource,
+)
 from .tasks import refresh_hq_datasource_task
 from .utils import (
-    ASYNC_DATASOURCE_IMPORT_LIMIT_IN_BYTES,
-    AsyncImportHelper,
     DomainSyncUtil,
-    download_datasource,
-    get_column_dtypes,
-    get_datasource_defn,
-    get_datasource_file,
     get_datasource_list_url,
     get_hq_database,
     get_schema_name_for_domain,
-    parse_date,
 )
+
+ASYNC_DATASOURCE_IMPORT_LIMIT_IN_BYTES = 5_000_000  # ~5MB
 
 logger = logging.getLogger(__name__)
 
@@ -110,40 +105,6 @@ class HQDatasourceView(BaseSupersetView):
         return redirect("/tablemodelview/list/")
 
 
-def convert_to_array(string_array):
-    """
-    Converts the string representation of a list to a list.
-    >>> convert_to_array("['hello', 'world']")
-    ['hello', 'world']
-
-    >>> convert_to_array("'hello', 'world'")
-    ['hello', 'world']
-
-    >>> convert_to_array("[None]")
-    []
-
-    >>> convert_to_array("hello, world")
-    []
-    """
-
-    def array_is_falsy(array_values):
-        return not array_values or array_values == [None]
-
-    try:
-        array_values = ast.literal_eval(string_array)
-    except ValueError:
-        return []
-
-    if isinstance(array_values, tuple):
-        array_values = list(array_values)
-
-    # Test for corner cases
-    if array_is_falsy(array_values):
-        return []
-
-    return array_values
-
-
 def trigger_datasource_refresh(domain, datasource_id, display_name):
     if AsyncImportHelper(domain, datasource_id).is_import_in_progress():
         flash(
@@ -160,11 +121,11 @@ def trigger_datasource_refresh(domain, datasource_id, display_name):
         provider, token, domain, datasource_id
     )
     if size < ASYNC_DATASOURCE_IMPORT_LIMIT_IN_BYTES:
-        response = refresh_hq_datasource(
+        refresh_hq_datasource(
             domain, datasource_id, display_name, path, datasource_defn, None
         )
         os.remove(path)
-        return response
+        return redirect("/tablemodelview/list/")
     else:
         limit_in_mb = int(ASYNC_DATASOURCE_IMPORT_LIMIT_IN_BYTES / 1000000)
         flash(
@@ -200,116 +161,6 @@ def queue_refresh_task(
         g.user.get_id(),
     ).task_id
     AsyncImportHelper(domain, datasource_id).mark_as_in_progress(task_id)
-    return redirect("/tablemodelview/list/")
-
-
-def refresh_hq_datasource(
-    domain,
-    datasource_id,
-    display_name,
-    file_path,
-    datasource_defn,
-    user_id=None,
-):
-    """
-    Pulls the data from CommCare HQ and creates/replaces the
-    corresponding Superset dataset
-    """
-    database = get_hq_database()
-    schema = get_schema_name_for_domain(domain)
-    csv_table = Table(table=datasource_id, schema=schema)
-    column_dtypes, date_columns, array_columns = get_column_dtypes(
-        datasource_defn
-    )
-
-    converters = {
-        column_name: convert_to_array for column_name in array_columns
-    }
-    # TODO: can we assume all array values will be of type TEXT?
-    sqlconverters = {
-        column_name: postgresql.ARRAY(sqlalchemy.types.TEXT)
-        for column_name in array_columns
-    }
-
-    def to_sql(df, replace=False):
-        database.db_engine_spec.df_to_sql(
-            database,
-            csv_table,
-            df,
-            to_sql_kwargs={
-                "if_exists": "replace" if replace else "append",
-                "dtype": sqlconverters,
-            },
-        )
-
-    try:
-        with get_datasource_file(file_path) as csv_file:
-
-            _iter = pd.read_csv(
-                chunksize=10000,
-                filepath_or_buffer=csv_file,
-                encoding="utf-8",
-                parse_dates=date_columns,
-                date_parser=parse_date,
-                keep_default_na=True,
-                dtype=column_dtypes,
-                converters=converters,
-                iterator=True,
-                low_memory=True,
-            )
-
-            to_sql(next(_iter), replace=True)
-
-            for df in _iter:
-                to_sql(df, replace=False)
-
-        # Connect table to the database that should be used for exploration.
-        # E.g. if hive was used to upload a csv, presto will be a better option
-        # to explore the table.
-        expore_database = database
-        explore_database_id = database.explore_database_id
-        if explore_database_id:
-            expore_database = (
-                db.session.query(Database)
-                .filter_by(id=explore_database_id)
-                .one_or_none()
-                or database
-            )
-
-        sqla_table = (
-            db.session.query(SqlaTable)
-            .filter_by(
-                table_name=datasource_id,
-                schema=csv_table.schema,
-                database_id=expore_database.id,
-            )
-            .one_or_none()
-        )
-        if sqla_table:
-            sqla_table.description = display_name
-            sqla_table.fetch_metadata()
-        if not sqla_table:
-            sqla_table = SqlaTable(table_name=datasource_id)
-            # Store display name from HQ into description since
-            #   sqla_table.table_name stores datasource_id
-            sqla_table.description = display_name
-            sqla_table.database = expore_database
-            sqla_table.database_id = database.id
-            if user_id:
-                user = superset.appbuilder.sm.get_user_by_id(user_id)
-            else:
-                user = g.user
-            sqla_table.owners = [user]
-            sqla_table.user_id = user.get_id()
-            sqla_table.schema = csv_table.schema
-            sqla_table.fetch_metadata()
-            db.session.add(sqla_table)
-        db.session.commit()
-    except Exception as ex:  # pylint: disable=broad-except
-        db.session.rollback()
-        raise ex
-
-    # superset.appbuilder.sm.add_permission_role(role, sqla_table.get_perm())
     return redirect("/tablemodelview/list/")
